requirements.txt
pandas
yfinance
numpy
colorama
google-generativeai